moved all questions to ques file to avoid clutter
